+++ src/views/space-board/SpaceBoard.scss
@@ -21,16 +21,6 @@
       display: flex;
       align-items: center;
       gap: var(--spacing-unit);
-      .status_filter {
-        position: relative;
-        &-menu {
-          top: calc(32px + 6px);
-          right: 0;
-          position: absolute;
-          background-color: var(--color-white);
-          line-height: 2;
-        }
-      }
     }
 
     &__btn {
