+++ src/views/user-projects/UserProjects.vue
@@ -15,32 +15,10 @@
         />
       </template>
       <template #right>
-        <div class="status_filter" v-click-away="close">
-          <BIMDataButton
-            class="user-projects__header__btn-sort"
-            fill
-            squared
-            icon
-            @click="toggle"
-          >
-            <BIMDataIconFilterList size="s" />
-          </BIMDataButton>
-          <transition name="fade">
-            <div v-show="isOpen" class="status_filter-menu p-y-12 p-x-18">
-              <div
-                v-for="status in computedStatuses"
-                :key="status"
-                class="status-filter-menu__item"
-              >
-                <BIMDataCheckbox
-                  :text="$t(`ProjectStatusBadge.${status}`)"
-                  :modelValue="selectedStatuses.includes(status)"
-                  @update:modelValue="(checked) => onStatusChange(status, checked)"
-                />
-              </div>
-            </div>
-          </transition>
-        </div>
+        <StatusFilterButton
+          :projects="userProjects"
+          @update:filteredProjects="filteredProjects = $event"
+        />
         <BIMDataButton
           class="user-projects__header__btn-sort"
           fill
@@ -72,57 +50,30 @@
 </template>
 
 <script>
-import { computed, ref } from "vue";
+import { ref } from "vue";
 import { useListFilter } from "../../composables/list-filter.js";
 import { useListSort } from "../../composables/list-sort.js";
 import { useToggle } from "../../composables/toggle.js";
 import { useStandardBreakpoints } from "../../composables/responsive.js";
 import { useProjects } from "../../state/projects.js";
-import { projectStatus } from "../../utils/projects.js";
 
 // Components
 import AppBreadcrumb from "../../components/specific/app/app-breadcrumb/AppBreadcrumb.vue";
 import ViewHeader from "../../components/specific/app/view-header/ViewHeader.vue";
 import ProjectCard from "../../components/specific/projects/project-card/ProjectCard.vue";
+import StatusFilterButton from "../../components/specific/projects/status-filter-button/StatusFilterButton.vue";
 
 export default {
   components: {
     AppBreadcrumb,
     ViewHeader,
     ProjectCard,
+    StatusFilterButton
   },
   setup() {
     const { isOpen, close, toggle } = useToggle();
     const { userProjects } = useProjects();
-
-    const selectedStatuses = ref([]);
-
-    const onStatusChange = (status, isChecked) => {
-      if (isChecked) {
-        if (!selectedStatuses.value.includes(status)) {
-          selectedStatuses.value.push(status);
-        }
-      } else {
-        selectedStatuses.value = selectedStatuses.value.filter((s) => s !== status);
-      }
-    };
-
-    const computedStatuses = computed(() => {
-      const statuses = new Set();
-      userProjects.value.forEach((project) => {
-        statuses.add(projectStatus(project));
-      });
-      return Array.from(statuses);
-    });
-
-    const filteredProjects = computed(() => {
-      return userProjects.value.filter((project) => {
-        if (selectedStatuses.value.length === 0) {
-          return true;
-        }
-        return selectedStatuses.value.includes(projectStatus(project));
-      });
-    });
+    const filteredProjects = ref(userProjects.value);
 
     const { filteredList: displayedProjects, searchText } = useListFilter(
       filteredProjects,
@@ -136,15 +87,14 @@ export default {
 
     return {
       // References
+      userProjects,
+      filteredProjects,
       projects: displayedProjects,
       searchText,
-      selectedStatuses,
-      computedStatuses,
       isOpen,
       // Methods
       close,
-      toggle, 
-      onStatusChange,
+      toggle,
       sortProjects,
       // Responsive breakpoints
       ...useStandardBreakpoints(),
